Refactoring and cleanup of pyclass
